+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v0.2.8]
 ### Changed
 - cleaned up shutdown logic for outbound sender [#205](https://github.com/xmidt-org/caduceus/pull/205)
 - added resetting queue depth and current workers gauges to outbound sender [#205](https://github.com/xmidt-org/caduceus/pull/205)
@@ -95,7 +97,8 @@ fixed build upload
 ### Added
 - Initial creation
 
-[Unreleased]: https://github.com/Comcast/caduceus/compare/v0.2.7...HEAD
+[Unreleased]: https://github.com/Comcast/caduceus/compare/v0.2.8...HEAD
+[v0.2.8]: https://github.com/Comcast/caduceus/compare/v0.2.7...v0.2.8
 [v0.2.7]: https://github.com/Comcast/caduceus/compare/v0.2.6...v0.2.7
 [v0.2.6]: https://github.com/Comcast/caduceus/compare/v0.2.5...v0.2.6
 [v0.2.5]: https://github.com/Comcast/caduceus/compare/v0.2.4...v0.2.5